+++ src/main.ts
@@ -47,6 +47,7 @@ import {
   maybeAddComment,
   maybePrefixPackage,
   getPropertyAccessor,
+  impFile,
 } from './utils';
 import { camelToSnake, capitalize, maybeSnakeToCamel } from './case';
 import {
@@ -196,7 +197,7 @@ export function generateFile(ctx: Context, fileDesc: FileDescriptorProto): [stri
         `);
 
         if (options.outputTypeRegistry) {
-          const messageTypeRegistry = imp('messageTypeRegistry@./typeRegistry');
+          const messageTypeRegistry = impFile(options, 'messageTypeRegistry@./typeRegistry');
 
           chunks.push(code`
             ${messageTypeRegistry}.set(${fullName}.$type, ${fullName});
@@ -207,7 +208,6 @@ export function generateFile(ctx: Context, fileDesc: FileDescriptorProto): [stri
     );
   }
 
-  let hasServerStreamingMethods = false;
   let hasStreamingMethods = false;
 
   visitServices(fileDesc, sourceInfo, (serviceDesc, sInfo) => {
@@ -247,12 +247,7 @@ export function generateFile(ctx: Context, fileDesc: FileDescriptorProto): [stri
             chunks.push(generateGrpcClientImpl(ctx, fileDesc, serviceDesc));
             chunks.push(generateGrpcServiceDesc(fileDesc, serviceDesc));
             serviceDesc.method.forEach((method) => {
-              if (!method.clientStreaming) {
-                chunks.push(generateGrpcMethodDesc(ctx, serviceDesc, method));
-              }
-              if (method.serverStreaming) {
-                hasServerStreamingMethods = true;
-              }
+              chunks.push(generateGrpcMethodDesc(ctx, serviceDesc, method));
             });
           }
         }
@@ -273,7 +268,7 @@ export function generateFile(ctx: Context, fileDesc: FileDescriptorProto): [stri
     if (options.outputClientImpl === true) {
       chunks.push(generateRpcType(ctx, hasStreamingMethods));
     } else if (options.outputClientImpl === 'grpc-web') {
-      chunks.push(addGrpcWebMisc(ctx, hasServerStreamingMethods));
+      chunks.push(addGrpcWebMisc(ctx, hasStreamingMethods));
     }
   }
 
@@ -335,8 +330,8 @@ export function makeUtils(options: Options): Utils {
 function makeLongUtils(options: Options, bytes: ReturnType<typeof makeByteUtils>) {
   // Regardless of which `forceLong` config option we're using, we always use
   // the `long` library to either represent or at least sanity-check 64-bit values
-  const util = imp('util@protobufjs/minimal');
-  const configure = imp('configure@protobufjs/minimal');
+  const util = impFile(options, 'util@protobufjs/minimal');
+  const configure = impFile(options, 'configure@protobufjs/minimal');
 
   // Before esModuleInterop, we had to use 'import * as Long from long` b/c long is
   // an `export =` module and exports only the Long constructor (which is callable).
@@ -808,6 +803,8 @@ function generateDecode(ctx: Context, fullName: string, messageDesc: DescriptorP
     createBase = code`Object.create(${createBase}) as ${fullName}`;
   }
 
+  const Reader = impFile(ctx.options, 'Reader@protobufjs/minimal');
+
   // create the basic function declaration
   chunks.push(code`
     decode(
@@ -943,13 +940,12 @@ function generateDecode(ctx: Context, fullName: string, messageDesc: DescriptorP
   return joinCode(chunks, { on: '\n' });
 }
 
-const Writer = imp('Writer@protobufjs/minimal');
-const Reader = imp('Reader@protobufjs/minimal');
-
 /** Creates a function to encode a message by loop overing the tags. */
 function generateEncode(ctx: Context, fullName: string, messageDesc: DescriptorProto): Code {
   const { options, utils, typeMap } = ctx;
   const chunks: Code[] = [];
+
+  const Writer = impFile(ctx.options, 'Writer@protobufjs/minimal');
 
   // create the basic function declaration
   chunks.push(code`
@@ -1688,7 +1684,11 @@ function generateWrap(ctx: Context, fullProtoTypeName: string, fieldNames: Struc
 
   if (isFieldMaskTypeName(fullProtoTypeName)) {
     chunks.push(code`wrap(paths: string[]): FieldMask {
-      return {paths: paths};
+      const result = createBaseFieldMask();
+
+      result.paths = paths;
+
+      return result;
     }`);
   }